update workflow status queue usage
