+++ refactor/dispatcher/app/core/utils.py
@@ -20,6 +20,7 @@
 
 """Utility functions for the Dispatcher microservice."""
 
+import os
 from datetime import datetime, timezone
 from io import BytesIO
 from multiprocessing import Queue as MPQ
@@ -28,6 +29,7 @@ from typing import Any, Dict, List
 import cloudpickle as pickle
 import requests
 from app.core.dispatcher_logger import logger
+from dotenv import load_dotenv
 
 from covalent._results_manager import Result
 from covalent._shared_files.context_managers import active_lattice_manager
@@ -42,6 +44,11 @@ from covalent._shared_files.defaults import (
     subscript_prefix,
 )
 from covalent._workflow.lattice import Lattice
+
+load_dotenv()
+
+
+BASE_URI = os.environ.get("BASE_URI")
 
 
 def is_empty(mp_queue: MPQ):
@@ -295,3 +302,10 @@ def update_result_and_ui(result_obj: Result, task_id: int) -> Dict[str, str]:
     resp_1 = send_result_object_to_result_service(result_obj)
     resp_2 = send_task_update_to_ui(dispatch_id=result_obj.dispatch_id, task_id=task_id)
     return {"update_result_response": resp_1, "update_ui_response": resp_2}
+
+
+def send_cancel_task_to_runner(dispatch_id: str, task_id: int):
+    response = requests.delete(f"{BASE_URI}/api/v0/workflow/{dispatch_id}/task/{task_id}/cancel")
+    response.raise_for_status()
+
+    return response.json()["cancelled_dispatch_id"], response.json()["cancelled_task_id"]
